Move createStyled to the separate file
